requirements.txt
abacusai
python-dotenv==1.0.0
pydantic>=2.0.0
